requirements.txt
Django>=3.0.0
django-admin-list-filter-dropdown>=1.0.3
django-filter>=2.4.0
djangorestframework>=3.0.0
drf-nested-routers>=0.92.0
drf-yasg2==1.19.3
python-frontmatter==0.5.0
inflection>=0.5.1